Move `HasMagic` trait impl codegen to its own file
